Update TdlpackIO.py
Updated method for indexing.  Now using a single dictionary for a file where the value of each key is a list with N-elements where N is the total number of records in the file.  This will allow for more efficient filtering/selection in a future commit.

NOTE:  Need to fix an issue with unpacking station id records...