Added file store dictionary
Added _open_file_store dictionary to track open files where key
is the filename and value is the open class object.

Added _source and _linked_station_record attrs for TdlpackRecord
object that will allow access to other records in the file object
mainly for accessing the station record when needed.